requirements.txt
mcp
snowflake-connector-python
toml
